requirements.txt
Django==2
djangorestframework==3.7.7
markdown
psycopg2
